try control port placement

+++ src/main/java/com/github/ilms49898723/fluigi/device/component/Port.java
@@ -24,7 +24,7 @@ public class Port extends BaseComponent {
         setHeight(2 * radius);
     }
 
-    public double getRadius() {
+    public int getRadius() {
         return mRadius;
     }
